Add elm-test and elm-verify-examples as local deps
Relying on global dependencies is error-prone, so here
they are added as local deps.
`elm-doc-test` has changed its name to `elm-verify-examples`
and updated the test format. Code and dependency has been
updated to reflect this.

Since we have `elm-test` and `elm-verify-examples` as `npm`
deps we can also use `npm` scripts to run them.